fix(front): #1830.19 fix crawling issues
Fixed crawling issues related to dynamically generated images:

- Since graphviz image generation is not completely dynamic,
(or at least semrush doesn't detect dynamic images)
a "base image" was added to be overwritten on each code change.
This image is locally generated, and will change everytime
a new parameter is added or overwritten on its source code

- Plantuml images behaviour is way too random on asciidoc
since it is intended to be used on asciidoctor, not asciidoc,
so the plantuml example was removed.

- Git now will recognize dynamically generated diagrams,
since graphviz images need a "base image" to build

- Minor fixes on spanish text

- Removed deprecated link example that was breaking the build

locally rendered changes:


Former-commit-id: 3d7fafa3c613da85893f83b64c0c419a5ec48f6c