refac(cross): #2986 remove latex support
- In order to be able to update pelican to python3,
asciidoc must be removed and asciidoctor must be
used. This implies stopping using asciidoc filters.
One of those filters is latex. For future stack implementations
we must think about a tool that directly offers latex
support instead of relying on a dependency of a dependency.
In the meantime, all contents using latex will
use the automatically generated images instead.


Former-commit-id: e9c13afd3448231e3b569c421b3229213f98b5af