integrates\feat(back): #2268.10 new xls generation method
- Now, openpyxls is replaced by pyexcelerate
- This is because the time performance is too much better
using this library. The performance chart shown in
https://pypi.org/project/PyExcelerate/
looks very sexy.
- Again, generation method is changed from rows ro ranges.
Now, each row is handled as a range of cells
- All styles are included via code, and pyexcelerate
doesn't provide too much customization of that, so report
style can be a bit more ugly, but everything is for a better
performance :)
- Next work will be on cleaning some code and replacing the
report template by a code solution, so this
`from openpyxl import load_workbook` will be removed.
ITReport class will be refactored. ITReport class will be
refactored.

Former-commit-id: 1dc21627cbc977d6f7f224cc696037941c2c25ba