integrates\feat(back): #2093.10 single vuln dataloader
- Current vulnerability dataloader returns a list of vulns
based on its finding_id, this is, returns the list of vulns of
a concrete finding. This cant be used in the new vulnerability
resolver.
- This new dataloader returns only one vulnerability based on
the provided uuid, having this way coherence with the recent
vuln_domain/get change. For now, this will be used just by
vulnerability resolver by passing only one uuid.
- This single_vulnerability dataloader will be renamed to just
vulnerability, and the actual vulnerability dataloader will be
called finding_vulnerabilities, having this way more coherence
with what those two do. This is an upcoming migration.


Former-commit-id: 24ad97db2e9455cdff6dce7b0dd04b0a0c6093c4